Homework 4:     Updated to new instructor repo. Run submit_quick_test.sh for minimal     working example of the weird time issue I am encountering where the     sequential version is significantly faster when executed for  the     second time.

+++ homework/4/src/parallelSpMV.c
@@ -6,8 +6,8 @@ void parallelMatrixConversion()
     int **top;
 
     //set number of row blocks and column blocks
-    nrowblks = ceil(numrows/(float)(block_width));
-    ncolblks = ceil(numcols/(float)(block_width));
+    nrowblks = ceil(numrows/(double)(block_width));
+    ncolblks = ceil(numcols/(double)(block_width));
 
     //allocate memory 
     parMatrixBlock = (parblock *)malloc(nrowblks*ncolblks * sizeof(parblock));
@@ -32,11 +32,11 @@ void parallelMatrixConversion()
     {
         k1 = colptrs[c];
         k2 = colptrs[c + 1] - 1;
-        blkc = ceil((c + 1)/(float)block_width);
+        blkc = ceil((c + 1)/(double)block_width);
         for(k = k1 - 1 ; k < k2 ; k++)
         {
             r = irem[k];
-            blkr = ceil(r/(float)block_width);
+            blkr = ceil(r/(double)block_width);
             parMatrixBlock[(blkr - 1) * ncolblks + (blkc - 1)].nnz++;
         }
     }
@@ -68,12 +68,12 @@ void parallelMatrixConversion()
     {
         k1 = colptrs[c];
         k2 = colptrs[c + 1] - 1;
-        blkc = ceil((c + 1)/(float)block_width);
+        blkc = ceil((c + 1)/(double)block_width);
 
         for(k = k1 - 1 ; k < k2 ; k++)
         {
             r = irem[k];
-            blkr = ceil(r/(float)block_width);
+            blkr = ceil(r/(double)block_width);
 
             parMatrixBlock[(blkr - 1) * ncolblks + blkc - 1].rloc[top[blkr - 1][blkc - 1]] = r - parMatrixBlock[(blkr - 1) * ncolblks + blkc - 1].roffset;
             parMatrixBlock[(blkr - 1) * ncolblks + blkc - 1].cloc[top[blkr - 1][blkc - 1]] = (c + 1) -  parMatrixBlock[(blkr - 1) * ncolblks + blkc - 1].coffset;
